// src/client/types.gen.ts
export type EventImagePublic = {
    id: string;
    event_id: string;
    image_url: string;
    display_order: number;
    created_at: string;
};

export type EventPublic = {
    title: string;
    short_summary?: (string | null);
    start_datetime: string;
    end_datetime: string;
    timezone?: string;
    recurring_series?: boolean;
    description?: (string | null);
    location?: (string | null);
    venue_name?: (string | null);
    latitude?: (number | null);
    longitude?: (number | null);
    show_on_explore?: boolean;
    password_protected?: boolean;
    event_activity_social_feed?: boolean;
    event_activity_updates_only?: boolean;
    youtube_video_url?: (string | null);
    flyer_url?: (string | null);
    spotify_song_url?: (string | null);
    title_font?: (string | null);
    accent_color?: (string | null);
    is_featured?: boolean;
    id: string;
    status: EventStatus;
    created_at: string;
    updated_at: string;
    published_at: (string | null);
    tickets?: Array<TicketPublic>;
    gallery_images?: Array<EventImagePublic>;
};
// ...
export type EventStatus = 'draft' | 'published' | 'cancelled' | 'completed';
// ...
export type TicketPublic = {
    id: string;
    event_id: string;
    name: string;
    description: (string | null);
    price_amount: number;
    price_currency: string;
    gross_price: (number | null);
    display_price: (number | null);
    quantity_total: number;
    quantity_sold: number;
    is_unlimited: boolean;
    sale_start: (string | null);
    sale_end: (string | null);
    validity_start: (string | null);
    validity_end: (string | null);
    min_per_order: number;
    max_per_order: number;
    sell_in_bundles: boolean;
    created_at: string;
    updated_at: string;
};